fix autostart compatibility
Commit 39c15e2 broke the startup script on webOS <7.

The Busybox timeout requires different arguments on old and new webOS.
Until webOS 6 (v1.29.3), the timeout duration must be specified with -t.
As of webOS 7 (v1.31.1), -t is no longer accepted.

This script tries to run timeout with -t first. If that fails, it checks
the exit status. A status of 143 indicates that the process was killed
with SIGTERM, which in this case (probably) means it timed out. If the
command failed for any other reason (such as not accepting -t), try it
again without -t.

This is a shorter version that puts the command in a variable.